api_gateway/requirements.txt
# api_gateway/requirements.txt
fastapi
uvicorn
sqlalchemy
aiohttp
python-multipart
python-jose
PyJWT
passlib
python-dotenv
pydantic
pydantic-settings
email-validator
boto3
psycopg2-binary
alembic
requests